source-salesforce-native: move authentication related code into auth.py
This is purely a re-organization of existing code to pre-position for a new authentication method. Keeping all the authenitcation related code in models.py would get unwieldy after the upcoming addition in the next commit. There are no funcitonality changes in this commit.
